[bugfix] [shared/oval, RHEL/6, RHEL/7, Fedora]
* Get rid of multiple not synced copies of idtranslate.py
  (the only left separate copy of idtranslate.py script is
  kept in Fedora/transforms, but we will deal with that one
  later since it needs / assumes another change yet),

* Also make common shared/testcheck_module.py to be called
  by each of numerous instances of testcheck.py Python script
  (used within various products) rather than keeping numerous
  separate copies of testcheck.py script itself (one for each
  of the products),

* Also add .gitignore for Fedora/input/checks & shared/ directories
  (git not to report *.pyc or testids.ini files for changes).
