--- Fedora/input/checks/idtranslate.py
@@ -1,1 +1,0 @@
-../../transforms/idtranslate.py
