+++ shared/.gitignore
@@ -1,0 +1,3 @@
+# files not to track in git
+*.pyc
+*.ini
